Change website style; add style for "recipe"
Add another stylesheet (style-recipe.css) to the website to accommodate
changes to website layout, as the rules added to a singular stylesheet
would be nonsensical to maintain and define.
Add minor changes to the layout to make reading easier
and more comprehendible.
Add changes to body, titles, headings, and a span highlighter class,
among other things.