Initiate a new style and label page sections
Initiate a new style because the older one was not friendly to adding
"separator elements" (div, span) and was not mobile to responsive
design. This is a primitive first step attempt into meeting those two
goals.

Pages are now organized into section blocks via dividers to reflect
the material.

Website now looks like an older, typewritten document for appeal.